refactor(03-05): clean-break config rewire — delete getters/machinery
- DELETE config/core/ (registry/provider/validator) + every schema.py +
  to_dict/from_dict machinery + the getters (get_config_registry,
  get_*_config_provider) — no vestigial compat shim (D-01)
- config/__init__.py: clean grouped re-export of Pydantic models + Settings +
  reference-data constants (mirrors core/enums style)
- split models into per-domain files (portfolio/trading/data/system/exchange.py);
  models.py re-exports them as the aggregate import home
- rewire consumers to construct models directly: itrader/__init__.py builds
  SystemConfig.default(); execution_handler uses SystemConfig.default();
  portfolio_handler holds a PortfolioConfig (drops provider + validate)
- exchange presets behavior-preserving (4 presets incl. low_latency) from the
  deleted presets.py; fix test import of old config.exchange.config path
- prune stale mypy overrides (legacy_config, config.exchange.schema deleted)
- config tree 3,380 -> 1,132 lines; mypy --strict clean; suite 321 pass;
  behavioral oracle byte-exact
