--- itrader/config/core/__init__.py
@@ -1,29 +1,0 @@
-"""
-Core configuration management components.
-
-This module provides the foundational components for configuration management:
-- ConfigRegistry: Central registry for domain configurations
-- ConfigProvider: Abstract provider interface and implementations
-- ConfigValidator: Validation framework for configuration data
-"""
-
-from .registry import ConfigRegistry
-from .provider import ConfigProvider, FileConfigProvider, RuntimeConfigProvider
-from .validator import ConfigValidator, SchemaValidator, BusinessValidator, ValidationError, ValidationResult
-
-__all__ = [
-    # Registry
-    'ConfigRegistry',
-    
-    # Providers
-    'ConfigProvider',
-    'FileConfigProvider', 
-    'RuntimeConfigProvider',
-    
-    # Validators
-    'ConfigValidator',
-    'SchemaValidator',
-    'BusinessValidator',
-    'ValidationError',
-    'ValidationResult'
-]
