--- itrader/config/data/__init__.py
@@ -1,24 +1,0 @@
-"""
-Data configuration domain.
-
-This module provides data-specific configuration management,
-including configuration classes for data sources, feeds, and storage.
-"""
-
-from .config import (
-    DataConfig, DataSource, DataFrequency, StorageType,
-    DataSourceConfig, DataFeedConfig, StorageConfig, ProcessingConfig, RealTimeConfig
-)
-
-__all__ = [
-    # Configuration classes
-    'DataConfig',
-    'DataSource',
-    'DataFrequency',
-    'StorageType',
-    'DataSourceConfig',
-    'DataFeedConfig',
-    'StorageConfig',
-    'ProcessingConfig',
-    'RealTimeConfig'
-]
