--- itrader/config/system/__init__.py
@@ -1,37 +1,0 @@
-"""
-System configuration domain.
-
-This module provides system-specific configuration management,
-including application settings, performance tuning, and logging configuration.
-"""
-
-from .config import (
-    SystemConfig, Environment, LogLevel,
-    PerformanceSettings, SecuritySettings, DatabaseSettings,
-    NotificationSettings, MonitoringSettings
-)
-from .logging_config import (
-    LoggingConfig, LogHandler, LogFormat, LogRotation,
-    get_default_logging_config, get_production_logging_config, get_debug_logging_config
-)
-
-__all__ = [
-    # Configuration classes
-    'SystemConfig',
-    'Environment',
-    'LogLevel',
-    'PerformanceSettings',
-    'SecuritySettings',
-    'DatabaseSettings',
-    'NotificationSettings',
-    'MonitoringSettings',
-    
-    # Logging configuration
-    'LoggingConfig',
-    'LogHandler',
-    'LogFormat',
-    'LogRotation',
-    'get_default_logging_config',
-    'get_production_logging_config',
-    'get_debug_logging_config'
-]
